refactor: move start list to module
