+++ src/app/app.component.html
@@ -1,8 +1,8 @@
 <h1>Zeitmessung</h1>
 <nav>
-  <button (click)="showStartList()">Startliste</button>
-  <button (click)="showTimekeeping()">Zeitmessung</button>
-  <button (click)="showEvaluation()">Auswertung</button>
+  <button (click)="displayTarget = 'startlist'">Startliste</button>
+  <button (click)="displayTarget = 'timekeeping'">Zeitmessung</button>
+  <button (click)="displayTarget = 'evaluation'">Auswertung</button>
 </nav>
 <app-start-list *ngIf="displayTarget === 'startlist'"></app-start-list>
 <app-timekeeping *ngIf="displayTarget === 'timekeeping'"></app-timekeeping>
