+++ src/app/start-list/participant.ts
@@ -1,0 +1,6 @@
+export interface Participant {
+  numberPlate: number;
+  name: string;
+  isSpare?: boolean;
+  category: string;
+}
