+++ src/app/start-list/start-list/start-list.component.css
